move some video cache logic to videoCache.js
